aegis-bridge/requirements.txt
# Aegis Bridge dependencies

# Supabase client
supabase>=2.0.0

# HTTP client for Archon72 API
httpx>=0.27.0

# Type hints
typing-extensions>=4.0.0

# Environment variable loading
python-dotenv>=1.0.0
